walmart-sales-pipeline/requirements.txt

snowflake-connector-python
pandas
matplotlib
seaborn
dbt-core
